TASK: Use anonymous class pattern instead of factory + class
Having a factories AND the handler lying around is just a bit too much, and might lead to confusion what to register in yaml.
